Refactor: /NewConference/ with /Modal/ (#80)
